machine: mentry: pass an embedded DTB if desired
This patch, based originally on a patch from Wesley Terpstra
<wesley@sifive.com>, bundles the DTB passed to the operating system
into the bbl.bin image itself.  This is important, since the
DTB that is supplied by the FSBL may be out of date with regards
to what the operating system kernel expects.

This feature is enabled by specifying the "--enable-dtb" switch
to configure.  The path to the DTB file can be supplied as an
argument to the "--with-dtb-file" switch.