machine: mentry: pass in a DTB from BBL compilation
This patch, almost all of which was originally written by Wesley
Terpstra <wesley@sifive.com>, bundles the DTB passed to Linux in with
the bbl.bin image itself.  This feature is important for our current
FPGA boot flow: it does not rely on an FSBL, and thus a new DTB can't
be passed in via the FSBL.

The DTB to be passed in should be named "linux.dtb" and should be
placed in the riscv-pk top-level directory.