machine: fdt: look for the compatible string used by Linux
The BBL code only looks at the first entry in an array of compatible
strings.  This causes problems with Linux-oriented DT files, since the
correct initial DT compatible entry will be "sifive,plic-1.0.0"
for the FU540 PLIC.  Take the easy way out and look for this instead
of "riscv,plic0".