machine: fdt: don't load CLINT and hart configuration from the DT
Linux doesn't currently use the CLINT, so there's no DT data for it.
This causes problems for the BBL, which expects to find CLINT data in
the DT.  We could modify the BBL to take two DT files, one for its
requirements and another for Linux's requirements, but it's easier
just to work around the issue by detecting the SoC model from the DT
SoC compatible string, and then use that to preload the CLINT details.
We also derive the hart data from the same source so we can get rid of
hart_phandles.  There's more work here to be done; the BBL can't
assume that it and Linux (or other operating system kernels) share the
same DT format.