machine: fdt: don't filter harts
Let whoever provided the DT to us decide which harts should be
filtered, rather than assuming it in this code and changing the DT
data.  The basic idea here is that the DT data that's passed to the
BBL to be passed to Linux should not be modified.  If the BBL really
needs its own DT file, it should get that separately.