machine: add the "hifive_unleashed" machine data configuration method
Add a "hifive_unleashed" machine data configuration method.  This can
be used when the output BBL image is explicitly targeted for the
SiFive HiFive Unleashed board with the FU540-C000 SoC.  The primary
virtue of this configuration method is that it does not rely on DT
parsing at all.  BBL will pass whatever DTB it was given to the
operating system kernel, without trying to parse or edit it.